refactor(tool): add package args for tool
